코드 컨벤션 점검 및 오류 수정

+++ src/main/java/com/hobbyhop/global/response/PageResponseDTO.java
@@ -12,34 +12,28 @@ import java.util.List;
 public class PageResponseDTO<E> {
     private int page;
     private int size;
-    private int total;
-
+    private long total;
     private int start;
     private int end;
-
     private boolean prev;
     private boolean next;
-
     private List<E> dtoList;
 
     @Builder(builderMethodName = "withAll")
-    public PageResponseDTO(PageRequestDTO pageRequestDTO, List<E> dtoList, int total) {
+    public PageResponseDTO(PageRequestDTO pageRequestDTO, List<E> dtoList, long total) {
         if(total <= 0) {
             return;
         }
+
         this.page = pageRequestDTO.getPage();
         this.size = pageRequestDTO.getSize();
-
         this.total = total;
         this.dtoList = dtoList;
-
-        //페이지 리스트 10개를 나타내기 위함.
 
         this.end = (int)Math.ceil((this.page / 10.0)) * 10;
         this.start = this.end - 9;
 
         int last = (int)(Math.ceil((total/(double)size)));
-
         this.end = end > last ? last: end;
         this.prev = this.start > 1;
         this.next = total > this.end * this.size;
